requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.4
Django==3.2.8
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.2
isort==5.10.1
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
zipp==3.7.0

mixer~=7.2.2